refactor: rename and extract quotas
