+++ Applications/AdminUi/packages/admin_api_types/lib/src/admin_api_types_base.dart
@@ -2,6 +2,5 @@ export 'clients/clients.dart';
 export 'identities/identities.dart';
 export 'messages/messages.dart';
 export 'metrics/metrics.dart';
-export 'quotas/quotas.dart';
 export 'relationships/relationships.dart';
 export 'tiers/tiers.dart';
