+++ Applications/AdminUi/packages/admin_api_types/lib/src/identities/identities.dart
@@ -1,3 +1,5 @@
 export 'identity.dart';
 export 'identity_deletion_process.dart';
 export 'identity_deletion_process_audit_log_entry.dart';
+export 'identity_quota.dart';
+export 'individual_quota.dart';
