--- Applications/AdminUi/packages/admin_api_types/lib/src/quotas/quotas.dart
@@ -1,1 +1,0 @@
-export 'quota.dart';
