+++ Applications/AdminUi/packages/admin_api_types/lib/src/tiers/tiers.dart
@@ -2,3 +2,4 @@ export 'create_tier_response.dart';
 export 'tier.dart';
 export 'tier_details.dart';
 export 'tier_overview_response.dart';
+export 'tier_quota_definition.dart';
